changes scripts to typescript
