+++ backend/scripts/initial_deploy.ts
@@ -1,14 +1,14 @@
-const hre = require("hardhat");
+import { ethers } from "hardhat";
 
 async function main() {
 // Member Manager
-const MemberManager = await hre.ethers.getContractFactory("MemberManager");
+const MemberManager = await ethers.getContractFactory("MemberManager");
 const memberManager = await MemberManager.deploy();
 await memberManager.deployed();
 console.log("Member Manager deployed to:", memberManager.address);
 
 // Proposal Manager
-const ProposalManager = await hre.ethers.getContractFactory("ProposalManager");
+const ProposalManager = await ethers.getContractFactory("ProposalManager");
 const proposalManager = await ProposalManager.deploy();
 await proposalManager.deployed();
 console.log("Proposal Manager deployed to:", proposalManager.address);
@@ -18,7 +18,7 @@ await memberManager.setProposalManager(proposalManager.address);
 await proposalManager.setMemberManager(memberManager.address);
 
 // Mastar DAO
-  const MasterDAO = await hre.ethers.getContractFactory("MasterDAO");
+  const MasterDAO = await ethers.getContractFactory("MasterDAO");
   const masterdao = await MasterDAO.deploy("https://github.com/realtakahashi/dao4.commons.shiden","Shin Takahashi",
     memberManager.address, proposalManager.address);
   await masterdao.deployed();
